Refactored out some common code for SubwayStatus and moved files.
