+++ assets/src/apps/v2/gl_eink.tsx
@@ -41,7 +41,7 @@ import OvernightDepartures from "Components/v2/eink/overnight_departures";
 import MultiScreenPage from "Components/v2/multi_screen_page";
 import SimulationScreenPage from "Components/v2/simulation_screen_page";
 import ExceptionCatcher from "Components/v2/exception_catcher";
-import SubwayStatus from "Components/v2/eink/subway_status";
+import SubwayStatus from "Components/v2/subway_status/eink/subway_status";
 
 const TYPE_TO_COMPONENT = {
   screen_normal: NormalScreen,
